% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{AgeToLengthComp}
\alias{AgeToLengthComp}
\title{Convert Age Composition to Length Composition}
\usage{
AgeToLengthComp(lh, S_a, tyears, N_at, comp_sample, sample_type = "catch")
}
\arguments{
\item{lh}{A data frame or list containing length-height relationship data.}

\item{S_a}{Matrix of selectivity at age.}

\item{tyears}{Total number of years in the data.}

\item{N_at}{Matrix of numbers at age over time.}

\item{comp_sample}{Vector of sample sizes for each year.}

\item{sample_type}{Type of sample, default is 'catch'. Can be 'catch' or other types.}
}
\value{
A list containing matrices: probabilities being in a length bin given age (plba),
probabilities being harvested at an age (page), probabilities of sampling a
given length bin (plb), and length frequencies (LF).
}
\description{
This function converts age composition data to length composition data based on
provided length-height (lh) relationships, selectivity (S_a), and other parameters.
}
