% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build Functions for FLjjm Package}
\description{
These functions are used to build various FLR (Fisheries Library in R) objects
from the output of the JJMS (Just Another Management Strategy) stock assessment
model. The functions handle different types of outputs including biological,
fishery, stock, and index data.
}
\details{
The primary functions include:
\itemize{
\item \code{buildFLBjjm}: Constructs an FLBiol object from jjm.output.
\item \code{buildFLBsjjm}: Builds multiple FLBiol objects from jjm.output.
\item \code{buildFLFsjjm}: Creates FLFisheries objects from jjm.output.
\item \code{buildFLIsjjm}: Generates FLIndices objects from jjm.output.
\item \code{buildFLRPsjjm}: Constructs FLPar objects with reference points from jjm.output.
\item \code{buildFLSojjm}: Builds FLStock objects from jjm.output considering areas.
\item \code{buildFLSjjm}: Constructs FLStock objects from jjm.output.
\item \code{buildFLSsjjm}: Creates multiple FLStock objects from jjm.output.
}

Each function takes a jjm.output object as input, which is the result of running
the JJMS model. They process this output into various FLR objects for further
analysis and use within the FLR framework.
}
\seealso{
\code{\link[FLCore]{FLBiol}}, \code{\link[FLFishery]{FLFisheries}},
\code{\link[FLCore]{FLIndices}}, \code{\link[FLCore]{FLStock}}
}
\author{
Iago Mosqueira (WMR) \email{iago.mosqueira@wur.nl}
}
