% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildFLBjjm}
\alias{buildFLBjjm}
\title{Build an FLBiol from a jjm.output object}
\usage{
buildFLBjjm(out, stock = 1, name = "CJM")
}
\arguments{
\item{out}{A \emph{jjm.output} object, as returned by readJJM.}

\item{stock}{Stock to extract, of relevance on 2-stock model runs, \emph{numeric}.}

\item{spwn}{Proportion of the year when spawning takes place.}
}
\value{
An object of class FLBiol.
}
\description{
Build an FLBiol from a jjm.output object
}
\seealso{
\link[FLCore:FLBiol]{FLCore::FLBiol}
}
