% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildFLFsjjm}
\alias{buildFLFsjjm}
\title{Build FLFisheries from JJMS Model Output}
\usage{
buildFLFsjjm(out, stock = 1)
}
\arguments{
\item{out}{JJMS model output object.}

\item{stock}{The stock number to extract fisheries data for, default is 1.}
}
\value{
An \code{FLFisheries} object constructed from the JJMS model output.
}
\description{
Constructs an \code{FLFisheries} object from a JJMS model output, processing data for
each fishery and creating corresponding \code{FLCatch} and \code{FLFishery} objects.
}
