% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jjm.R
\name{buildjjmctl}
\alias{buildjjmctl}
\title{Updating dat and ctl file lists from FLstock(s) and FLIndices}
\usage{
buildjjmctl(stk, idx, dat, ctl, ...)
}
\arguments{
\item{stk}{\link{TODO:description}}

\item{idx}{\link{TODO:description}}

\item{dat}{\link{TODO:description}}

\item{ctl}{\link{TODO:description}}
}
\value{
\link{TODO:description}
}
\description{
Updating dat and ctl file lists from FLstock(s) and FLIndices
}
\examples{
data(cjmstk)
ctl <- buildjjmctl(stk, idx, mod$data, mod$control)
data(cjmstks)
ctl <- buildjjmctl(stks, idxs, mods$data, mods$control)
}
