% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{catch,FLStocks-method}
\alias{catch,FLStocks-method}
\title{Calculate Total Catch across FLStocks}
\usage{
\S4method{catch}{FLStocks}(object)
}
\arguments{
\item{object}{An FLStocks object.}
}
\value{
A numeric value representing the total catch.
}
\description{
This method calculates the total catch by summing the catch (accounting for
area sums) of each FLStock object within an FLStocks object.
}
