% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{cjm2.oem}
\alias{cjm2.oem}
\title{Update Observations in Fishery Operation Model}
\usage{
cjm2.oem(om, deviances, observations, args, tracking)
}
\arguments{
\item{om}{Fishery operation model object.}

\item{deviances}{Deviances for the model.}

\item{observations}{Observations data.}

\item{args}{Argument list containing model parameters.}

\item{tracking}{Tracking information.}
}
\value{
A list with updated stock data, indices, observations, and tracking information.
}
\description{
This function updates stock observations and indices in a fishery operation model
(OM) based on provided arguments and tracking information.
}
