% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{cjmage2len}
\alias{cjmage2len}
\title{Convert Age Composition to Length Composition for Landings}
\usage{
cjmage2len(
  landings,
  selex,
  ess = 100,
  L_inf = 80.4,
  k = 0.16,
  L_0 = 18,
  M = 0.33,
  CVlen = 0.09,
  ages = an(dimnames(landings)$age),
  sample_type = "catch"
)
}
\arguments{
\item{landings}{Matrix of landings at age.}

\item{selex}{Matrix of selectivity at age.}

\item{ess}{Effective sample size for each year, default is 100.}

\item{L_inf}{Asymptotic length, default is 80.4.}

\item{k}{Growth coefficient, default is 0.16.}

\item{L_0}{Theoretical length at age zero, default is 18.}

\item{M}{Natural mortality rate, default is 0.33.}

\item{CVlen}{Coefficient of variation in length, default is 0.09.}

\item{ages}{Vector of ages, default is derived from landings.}

\item{sample_type}{Type of sampling, default is 'catch'.}
}
\value{
A matrix representing length composition for each year.
}
\description{
This function converts age composition data from landings to length composition
using specified biological parameters and selectivity.
}
