% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{cjmfwc}
\alias{cjmfwc}
\title{Converts FLQuants with Fleets as Areas into a fwdControl Object}
\usage{
cjmfwc(flqs, quant = "catch", nstocks = 1)
}
\arguments{
\item{flqs}{An \code{FLQuants} object as returned by a HCR module.}

\item{quant}{The quant to use, defaults to 'catch'.}

\item{nstocks}{Number of stocks, defaults to 1.}
}
\value{
A \code{fwdControl} object with the corresponding FCB slot.
}
\description{
Converts \code{FLQuants} objects, typically returned by a Harvest Control Rule (HCR) module,
into a \code{fwdControl} object with different behavior based on the number of stocks.
}
