% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{cperfect.oem}
\alias{cperfect.oem}
\title{Create a Perfect Observation Error Model}
\usage{
cperfect.oem(
  stk,
  deviances,
  observations,
  args,
  tracking,
  biomass = FALSE,
  ...
)
}
\arguments{
\item{stk}{Stock data.}

\item{deviances}{Deviances for the model.}

\item{observations}{Observations data.}

\item{args}{Argument list.}

\item{tracking}{Tracking information.}

\item{biomass}{Boolean, if TRUE, use biomass instead of numbers.}

\item{...}{Additional arguments.}
}
\value{
A list with updated stock data, index, observations, and tracking.
}
\description{
This function creates a perfect observation error model for a stock assessment,
adjusting the stock data to match observations.
}
