% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.combinejjmsout}
\alias{.combinejjmsout}
\title{Combine Fisheries Model Results}
\usage{
.combinejjmsout(x, y)
}
\arguments{
\item{x}{First set of fisheries model results.}

\item{y}{Second set of fisheries model results.}
}
\value{
A combined list of fisheries model results.
}
\description{
Combines results from two sources of fisheries model runs, handling both single and
multiple stock scenarios. It combines data such as stock numbers, harvest, data, and control.
}
