% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{exejjms}
\alias{exejjms}
\title{Execute JJMS Model}
\usage{
exejjms(name, path, args = "", verbose = TRUE, clean = TRUE)
}
\arguments{
\item{name}{The name of the JJMS model.}

\item{path}{The path to the directory where the JJMS model is located.}

\item{args}{Additional arguments for the JJMS model execution.}

\item{verbose}{If TRUE, prints the execution log to the console.}

\item{clean}{If TRUE, cleans up the output directory after model execution.}
}
\value{
Invisible path where the model was executed.
}
\description{
Executes the JJMS model with specified arguments, manages file operations for the model run,
and optionally cleans up the output directory.
}
