% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{fbar,FLStocks-method}
\alias{fbar,FLStocks-method}
\title{Calculate Average Fishing Mortality across FLStocks}
\usage{
\S4method{fbar}{FLStocks}(object)
}
\arguments{
\item{object}{An FLStocks object.}
}
\value{
A numeric value representing the average fishing mortality.
}
\description{
This method calculates the average fishing mortality (fbar) across multiple
FLStock objects contained within an FLStocks object.
}
