% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLjjm.R
\name{FLjjm}
\alias{FLjjm}
\title{FLjjm: An R Package for Interacting with JJMS Stock Assessment Model in FLR}
\description{
This package provides tools for the Fishery Library (FLR) to interact with the
JJMS (Just Another Management Strategy) stock assessment model. The package
includes functions for building, reading, and executing JJMS assessments
within the FLR framework.
}
\details{
The main functions include:
\itemize{
\item \code{buildFLjjm}: Function to build FLjjm objects.
\item \code{readFLjjm}: Function to read FLjjm data from files.
\item \code{loadJJMS}: Function to load JJMS data.
\item \code{exejjms}: Function to execute the JJMS model.
\item \code{runjjms}: Function to run JJMS simulations.
\item \code{jjms}: General function for JJMS model operations.
}

The package aims to facilitate the integration of JJMS model outputs with the FLR
tools for fisheries science and management strategy evaluation.
}
\seealso{
\code{\link[vignette]{FLjjm}} in the FLjjm package for an
overview of the package.
}
\author{
Iago Mosqueira (WMR) \email{iago.mosqueira@wur.nl}

Karolina Molla Gazi (WMR) \email{karolina.mollagazi@wur.nl}
}
