% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{fwdmov.om}
\alias{fwdmov.om}
\title{Forward Movement Operation Model}
\usage{
fwdmov.om(om, ctrl, FCB = FCB(ctrl), rates, time = 0, ...)
}
\arguments{
\item{om}{A fishery operation model object.}

\item{ctrl}{Control measures data frame or matrix.}

\item{FCB}{Fishing mortality, default is derived from \code{ctrl}.}

\item{rates}{Movement rates matrix.}

\item{time}{Time step for the simulation, default is 0.}

\item{...}{Additional arguments.}
}
\value{
A list containing the updated fishery operation model object.
}
\description{
Simulates the movement of fish populations between two areas in a fishery
operation model. Adjusts population numbers based on movement rates, fishing
mortality, and natural mortality.
}
