% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{jjms.sa}
\alias{jjms.sa}
\title{JJMS Stock Assessment}
\usage{
jjms.sa(stk, idx, args, tracking, ...)
}
\arguments{
\item{stk}{Stock data.}

\item{idx}{Index data.}

\item{args}{Argument list.}

\item{tracking}{Tracking information.}

\item{...}{Additional arguments.}
}
\value{
A list with the resulting stock data, tracking information, and arguments.
}
\description{
Performs a JJMS stock assessment, modifying and running the JJMS model with provided arguments.
}
