% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadFLSjjms}
\alias{loadFLSjjms}
\title{Load FLStocks from JJMS Model}
\usage{
loadFLSjjms(path, combine = FALSE)
}
\arguments{
\item{path}{The path to the directory containing the JJMS model data.}

\item{combine}{Boolean, if TRUE, combine the loaded FLStocks into a single object.}
}
\value{
FLStocks object(s) loaded from the specified path.
}
\description{
This function loads FLStocks from JJMS model directories located at a specified path.
It can handle both single and multiple stock scenarios and offers an option to combine
the loaded stocks into a single object.
}
