% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadJJMS}
\alias{loadJJMS}
\title{Load JJMS Model Components}
\usage{
loadJJMS(name, path)
}
\arguments{
\item{name}{The name of the JJMS model.}

\item{path}{The path to the directory containing the JJMS model data.}
}
\value{
A list containing biological data, fisheries data, indices, reference points,
stock data, and model data and control settings.
}
\description{
Loads a JJMS model from a given path and constructs various components of a fisheries
model, including biological data, fisheries data, indices, reference points, and stock data.
}
