% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{namejjms}
\alias{namejjms}
\title{Get Name of JJMS Model}
\usage{
namejjms(path)
}
\arguments{
\item{path}{The path to the directory containing the JJMS model data.}
}
\value{
A string with the name of the JJMS model.
}
\description{
Reads the first configuration file in a JJMS model directory to extract the model's name.
}
