% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{nStocksjjms}
\alias{nStocksjjms}
\title{Get Number of Stocks in JJMS Model}
\usage{
nStocksjjms(name, path)
}
\arguments{
\item{name}{The name of the JJMS model.}

\item{path}{The path to the directory containing the JJMS model data.}
}
\value{
Numeric value representing the number of stocks in the JJMS model.
}
\description{
Reads a JJMS model control file to determine the number of stocks.
}
