% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{packjjmsrun}
\alias{packjjmsrun}
\title{Clean Up JJMS Model Run Directory}
\usage{
packjjmsrun(path)
}
\arguments{
\item{path}{The path to the JJMS model run directory.}
}
\value{
Invisible TRUE if all files are successfully removed, FALSE otherwise.
}
\description{
Removes unnecessary files from a JJMS model run directory, keeping only essential files.
}
