% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{perfcjm.sa}
\alias{perfcjm.sa}
\title{Perfect CJM Stock Assessment}
\usage{
perfcjm.sa(stk, idx, args, tracking, ...)
}
\arguments{
\item{stk}{Stock data.}

\item{idx}{Index data.}

\item{args}{Argument list.}

\item{tracking}{Tracking information.}

\item{...}{Additional arguments.}
}
\value{
A list with stock data, tracking information, and arguments.
}
\description{
A placeholder function for perfect CJM stock assessment, potentially for testing or
method development purposes.
}
