% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_roxygen_all.R, R/read.R
\name{readFLBjjm}
\alias{readFLBjjm}
\title{Create an FLBiol from a JJMS model run directory}
\usage{
readFLBjjm(name, path, stock = 1)
}
\arguments{
\item{name}{Name of the ctl model file, \emph{character}.}

\item{path}{Path ot the model folder structure, \emph{character}.}

\item{stock}{Numeric. Stock index to extract when the run contains multiple
stocks. Default is 1.}

\item{stock=1}{Stock to extract, of relevance on 2 stocks model runs, \emph{numeric}.}
}
\value{
An object of class \code{FLBiol} constructed from the jjm output
(with slots such as \code{n}, \code{wt}, \code{m}, \code{mat} filled
according to the model input and output files).

An object of class FLBiol.
}
\description{
An object of class \code{FLBiol} is created from the information in the
various files in a \code{jjms} folder structure. The function reads the
model run (control, input and results folders) and constructs the
biological component that can be used within the FLR framework. See
\code{buildFLBjjm} for details about how each slot is populated.

An object of class \emph{FLBiol} is created from the information in the various
files in a \emph{jjms} folder structure. See \emph{buildFLBjjm} for how each slot is
populated from the inputs and outputs of \emph{jjms}.
}
\details{
The function delegates the heavy lifting to \code{jjmR::readJJM}
to import the model run and then to \code{buildFLBjjm} which maps the
jjm inputs/outputs to the corresponding FLBiol slots. Typical jjm
outputs used include population numbers by age (N), weight-at-age,
natural mortality (M) and maturity. When a multi-stock jjm run is used,
specify \code{stock} to select the correct output component.
}
\examples{
\dontrun{
biol <- readFLBjjm(name = "h1_1.07",
  path = system.file("ext-data", "single_stock", package = "FLjjm"))
summary(biol)
}
bio <- readFLBjjm(name="h1_1.07",
  path=system.file("ext-data", "single_stock", package="FLjjm"))
summary(bio)
}
\references{
EUPL 1.2
}
\seealso{
\code{buildFLBjjm}, \code{jjmR::readJJM}

\link[jjmR:readJJM]{jjmR::readJJM} \link[FLCore:FLBiol]{FLCore::FLBiol} buildFLBjjm
}
\author{
Iago Mosqueira \href{mailto:iago.mosqueira@wur.nl}{iago.mosqueira@wur.nl}
}
