% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_roxygen_all.R, R/read.R
\name{readFLFsjjm}
\alias{readFLFsjjm}
\title{Create FLFisheries objects from a JJMS model run directory}
\usage{
readFLFsjjm(name, path)
}
\arguments{
\item{name}{Name of the ctl model file, \emph{character}.}

\item{path}{Path ot the model folder structure, \emph{character}.}

\item{stock}{Numeric. Stock index to extract when more than one stock is
present. Default is 1.}
}
\value{
A list (or an \code{FLFisheries}/\code{FLFishery} object) with
fisheries built for the given stock.

An object of class FLIndices.
}
\description{
Construct the \code{FLFisheries} (or individual \code{FLFishery}) objects
representing fisheries in the jjms output. For each fishery the function
creates a \code{FLCatch} with slots \code{landings}, \code{landings.n},
\code{landings.wt}, discards and related weights, using the model outputs
C_fsh_* and wt_fsh_*.

An object of class \emph{FLFisheries}, a list of \emph{FLFishery} objects, is created
from the information in the various files in a \emph{jjms} folder structure.
See \emph{buildFLFsjjm} for how each slot is populated from the inputs and
outputs of \emph{jjms}.
}
\details{
The function extracts fishery names from \code{out$output[[stock]]$Fshry_names}
and loops through the fisheries building \code{FLQuant} objects for
landings numbers (C_fsh_i), weights at age (wt_fsh_i) and other slots.
}
\examples{
\dontrun{
fis <- readFLFsjjm(name = "h1_1.07",
  path = system.file("ext-data", "single_stock", package = "FLjjm"))
str(fis)
}
fisheries <- readFLFsjjm(name="h1_1.07",
  path = system.file("ext-data", "single_stock", package="FLjjm"))
summary(fisheries)
}
\seealso{
\code{buildFLFsjjm}

\link[jjmR:readJJM]{jjmR::readJJM} \link[FLFishery:FLFisheries]{FLFishery::FLFisheries} \link[FLFishery:FLFishery]{FLFishery::FLFishery} buildFLFsjjm
}
\author{
Iago Mosqueira \href{mailto:iago.mosqueira@wur.nl}{iago.mosqueira@wur.nl}
}
