% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_roxygen_all.R, R/read.R
\name{readFLIsjjm}
\alias{readFLIsjjm}
\title{Create an FLIndices object from a JJMS model run directory}
\usage{
readFLIsjjm(name, path)
}
\arguments{
\item{name}{Name of the ctl model file, \emph{character}.}

\item{path}{Path ot the model folder structure, \emph{character}.}
}
\value{
An \code{FLIndices} object with one or more \code{FLIndex}
elements representing survey/observation time series extracted from the
model output.

An object of class FLIndices.
}
\description{
Read jjms output and construct an \code{FLIndices} object containing the
abundance indices used/produced by the model. The function maps index
selectivity patterns, catchability (q) and range/start-end fractions to
FLIndex objects and combines them into an FLIndices list.

An object of class \emph{FLIndices}, a list of \emph{FLIndex} objects, is created
from the information in the various files in a \emph{jjms} folder structure.
See \emph{buildFLIsjjm} for how each slot is populated from the inputs and
outputs of \emph{jjms}.
}
\details{
The function calls \code{jjmR::readJJM} to read the run and
\code{buildFLIsjjm} to construct the FLIndices object. The mapping
follows the conventions used in the package: selectivity patterns are
read from \code{output@sel_ind_i}, index catchability from
\code{output@q_i} and month ranges are converted to fractions using the
\code{Imonths} input.
}
\examples{
\dontrun{
idx <- readFLIsjjm(name = "h1_1.07",
  path = system.file("ext-data", "single_stock", package = "FLjjm"))
plot(idx)
}
indices <- readFLIsjjm(name="h1_1.07",
  path = system.file("ext-data", "single_stock", package="FLjjm"))
summary(indices)
}
\seealso{
\code{buildFLIsjjm}, \code{jjmR::readJJM}

\link[jjmR:readJJM]{jjmR::readJJM} \link[FLCore:FLIndices]{FLCore::FLIndices} \link[FLCore:FLIndex]{FLCore::FLIndex} buildFLIsjjm
}
\author{
Iago Mosqueira \href{mailto:iago.mosqueira@wur.nl}{iago.mosqueira@wur.nl}
}
