% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_roxygen_all.R, R/read.R
\name{readFLoemjjm}
\alias{readFLoemjjm}
\title{Read FLIndices with OEM iterations and MC outputs}
\usage{
readFLoemjjm(name, path, method = cjm.oem, iter = NULL, ...)
}
\arguments{
\item{name}{Name of the ctl model file, \emph{character}.}

\item{path}{Path to the model folder structure, \emph{character}.}

\item{method}{Function. The OEM sampling function to be used; default is
\code{cjm.oem}.}

\item{iter}{Integer. Number of iterations to return/propgate. If NULL,
defaults to 1.}

\item{...}{Additional arguments passed to \code{method}.}
}
\value{
A list of \code{FLIndex} objects propagated to \code{iter}
iterations and augmented with MC-derived catchabilities when available.

An object of class FLoem.
}
\description{
Read jjm output and construct indices suitable for OEM simulations.
When \code{iter > 1} the function will attempt to read Monte Carlo
evaluation output to populate index catchability (q) and other stochastic
elements. The function returns indices propagated to the requested
number of iterations.

An object of class \emph{FLoem} is created from the information in the various
files in a \emph{jjms} folder structure. The \emph{@observations} slot is created
containing the past observations of catch and biology (as an \emph{FLStock} or
\emph{FLStocks}) and for the indices of abundance (as an \emph{FLIndices})
}
\examples{
\dontrun{
idx <- readFLoemjjm("h1_1.07", path = system.file("ext-data","single_stock", package="FLjjm"), iter = 10)
}
# One stock OM
oem <- readFLoemjjm(name="h1_1.07",
  path=system.file("ext-data", "single_stock", package="FLjjm"))
oem
# Two stock OM
oemtwo <- readFLoemjjm(name="h2_1.07",
  path=system.file("ext-data", "two_stock", package="FLjjm"))
oemtwo
}
\seealso{
\code{readFLIsjjm}, \code{readMCeval}

\link[jjmR:readJJM]{jjmR::readJJM} \link[mse:FLoem-class]{mse::FLoem} buildFLSjjm buildFLIsjjm
}
\author{
Iago Mosqueira \href{mailto:iago.mosqueira@wur.nl}{iago.mosqueira@wur.nl}
}
