% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_roxygen_all.R, R/read.R
\name{readFLomjjm}
\alias{readFLomjjm}
\title{Create an FLombf object from a JJMS model run directory}
\usage{
readFLomjjm(name, path, iter = NULL, ...)
}
\arguments{
\item{name}{Name of the ctl model file, \emph{character}.}

\item{path}{Path to the model folder structure, \emph{character}.}
}
\value{
An \code{FLombf} object ready to be used as an operating model
within the FLR/MSE framework.

An object of class FLBiol.
}
\description{
Construct an \code{FLombf} (operating model for biomass/forces framework)
from the outputs of a jjms run. The resulting object contains three main
slots populated by the corresponding \code{buildFL*jjm} helper functions:
\code{biols} (from \code{buildFLBjjm}), \code{fisheries} (from
\code{buildFLFsjjm}) and \code{refpts} (from \code{buildFLRPsjjm}).

An object of class \emph{FLombf} is created from the information in the various
files in a \emph{jjms} folder structure. Three slots are created from the
corresponding calls to \emph{buildFLjjm} functions.
\itemize{
\item @biols, of class \emph{FLBiols} from a call to \emph{buildFLBsjjm}.
\item @fisheries, of class \emph{FLFisheries}, from a call to \emph{buildFLFsjjm}.
\item @refpts, of classs \emph{FLPars}, from a call to \emph{buildFLRPsjjm}.
}
}
\details{
The wrapper reads the jjm run with \code{jjmR::readJJM} and then
assembles the subcomponents using the specialized builder functions.
The function is useful to convert an existing jjms assessment run into
an FLR-compatible operating model for simulation or MSE work.
}
\examples{
\dontrun{
om <- readFLomjjm(name = "h1_1.07",
  path = system.file("ext-data", "single_stock", package = "FLjjm"))
summary(om)
}
# One stock OM
om <- readFLomjjm(name="h1_1.07",
  path=system.file("ext-data", "single_stock", package="FLjjm"))
summary(om)
# Two stock OM
omtwo <- readFLomjjm(name="h2_1.07",
  path=system.file("ext-data", "two_stock", package="FLjjm"))
summary(omtwo)
}
\seealso{
\code{buildFLBjjm}, \code{buildFLFsjjm}, \code{buildFLRPsjjm}

\link[jjmR:readJJM]{jjmR::readJJM} FLombf buildFLBjjm buildFLFsjjm
}
\author{
Iago Mosqueira \href{mailto:iago.mosqueira@wur.nl}{iago.mosqueira@wur.nl}
}
