% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readFLSjjm}
\alias{readFLSjjm}
\title{Create an FLStock from a JJMS model run directory}
\usage{
readFLSjjm(name, path, stock = 1, output = TRUE)
}
\arguments{
\item{name}{Name of the ctl model file, \emph{character}.}

\item{path}{Path ot the model folder structure, \emph{character}.}

\item{stock=1}{Stock to extract, of relevance on 2 stocks model runs, \emph{numeric}.}
}
\value{
An object of class FLStock.
}
\description{
An object of class \emph{FLStock} is created from the information in the various
files in a \emph{jjms} folder structure. See \emph{buildFLSjjm} for how each slot is
populated from the inputs and outputs of \emph{jjms}.
}
\examples{
bio <- readFLSjjm(name="mod1.00",
  path = system.file("ext-data", "single_stock", package="FLjjm"))
summary(bio)
}
\seealso{
\link[jjmR:readJJM]{jjmR::readJJM} \link[FLCore:FLStock]{FLCore::FLStock} buildFLSjjm
}
