% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readFLSsjjm}
\alias{readFLSsjjm}
\title{Create an FLStocks from a two-stock JJMS model run directory}
\usage{
readFLSsjjm(name, path, output = FALSE)
}
\arguments{
\item{name}{Name of the ctl model file, \emph{character}.}

\item{path}{Path ot the model folder structure, \emph{character}.}
}
\value{
An object of class FLStock.
}
\description{
An object of class \emph{FLStocks} is created from the information in the various
files in a \emph{jjms} folder structure. See \emph{buildFLSjjm} for how each slot is
populated from the inputs and outputs of \emph{jjms}.
}
\examples{
bio <- readFLSsjjm(name="h2_1.07",
  path=system.file("ext-data", "two_stock", package="FLjjm"), output=TRUE)
summary(bio)
}
\seealso{
\link[jjmR:readJJM]{jjmR::readJJM} \link[FLCore:FLStock]{FLCore::FLStock} buildFLSjjm
}
