% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runjjms}
\alias{runjjms}
\title{Run JJMS Model}
\usage{
runjjms(mod, path = tempfile(), args = "", verbose = TRUE)
}
\arguments{
\item{mod}{The JJMS model data to be executed.}

\item{path}{The path to save model files and run the JJMS model, defaults to a temporary file path.}

\item{args}{Additional arguments for the JJMS model execution.}

\item{verbose}{If TRUE, execution details are printed; otherwise, they are suppressed.}
}
\value{
The path where the JJMS model was executed.
}
\description{
Executes the JJMS model with the given model data, saving the model files to a specified path.
The function creates necessary directories, writes model files, and calls the JJMS model.
}
