% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slick.R
\name{slick_performance}
\alias{slick_performance}
\title{Compute performance metrics for MPs and OM in slick format}
\usage{
slick_performance(
  x,
  om,
  statistics,
  metrics = list(C = catch, SB = ssb, F = fbar)
)
}
\arguments{
\item{x}{A list of MP or OM projections runs, of class 'list' or 'FLmse'.}

\item{om}{An OM, of class 'FLom' or 'FLombf'.}

\item{statistics}{A list of performance statistics to compute, 'list'.}

\item{metrics}{Metrics to compute, 'list'.}
}
\value{
A combined data.table with performance metrics
}
\description{
This function computes performance metrics for MPs over future years and for OM over past years.
}
\examples{
# slick_performance(x = my_mps_data, om = my_om_data, statistics = my_statistics)

}
