% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{ssb,FLStocks-method}
\alias{ssb,FLStocks-method}
\title{Calculate Total Spawning Stock Biomass across FLStocks}
\usage{
\S4method{ssb}{FLStocks}(object)
}
\arguments{
\item{object}{An FLStocks object.}
}
\value{
A numeric value representing the total spawning stock biomass.
}
\description{
This method calculates the total spawning stock biomass (ssb) by summing
the ssb of each FLStock object within an FLStocks object.
}
