% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{yourFunctionName}
\alias{yourFunctionName}
\title{Construct Components of Fisheries Model}
\usage{
yourFunctionName(mod)
}
\arguments{
\item{mod}{The model object from which components are to be constructed.}
}
\value{
A list containing biological data, fisheries data, indices, reference points,
stock data, and model data and control settings.
}
\description{
This function constructs various components of a fisheries model from a given model object.
It builds biological data, fisheries data, indices, reference points, and stock data.
}
