% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildFLIsjjm}
\alias{buildFLIsjjm}
\title{Build an FLIndices object from a jjm.output list}
\usage{
buildFLIsjjm(out)
}
\arguments{
\item{out}{A one or two-stock, single run, jjm.output object.}

\item{stock=1}{Stock for which indices are to be loaded.}
}
\value{
An \code{FLIndices} object.
}
\description{
This function creates an \code{FLIndices}, a list of \code{FLIndexBiomass} objects,
each of them containing the data and estimates of a single index of abundance.
Slots on each \code{FLIndexBiomass} object are filled with the following information:
\itemize{
\item index: \code{data$Index}.
\item index.var: \code{data$Indexerr}.
\item catch.n: \code{data$Ipropage}.
\item catch.wt: \code{data$Iwtatage}.
\item effort: Empty.
\item sel.pattern: \code{output@sel_ind_i}, where \code{i} refers to the index number.
\item index.q: \code{output@q_i}, where \code{i} refers to the index number.
\item range 'startf' and 'endf': (\code{data$Imonths} - 1) / 12.
}
}
\examples{
idx <- readFLIsjjm("mod1.00",
  path = system.file("ext-data", "single_stock", package="FLjjm"))
summary(idx)
plot(idx)
}
