% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildFLRPsjjm}
\alias{buildFLRPsjjm}
\title{Construct an FLPar contaning reference points from a jjm.output object}
\usage{
buildFLRPsjjm(out, stock = 1)
}
\arguments{
\item{name}{Name of the ctl model file, \emph{character}.}

\item{path}{Path ot the model folder structure, \emph{character}.}

\item{stock=1}{Stock to extract, of relevance on 2 stocks model runs, \emph{numeric}.}
}
\value{
An object of class FLPar.
}
\description{
An object of class \emph{FLPar} is created from the information in a jjm.output
object and for a given stock, if more than one is present.
The function currently extracts the following reference points from the
\code{msy_mt} data.frame in the \code{output} element of \code{jjm.output}:
\itemize{
\item MSY: Maximum Sustainable Yield, in thousands of tonnes.
\item SB0: Virgin spawning biomass, in thousands of tonnes.
\item SBmsy: Spawning biomass at MSY, in thousands of tonnes.
\item Fmsy: Fishing mortality at MSY.
}
}
\examples{
rps <- readFLRPsjjm(name = "mod1.00",
  path = system.file("ext-data", "single_stock", package="FLjjm"))
summary(rps)
}
\seealso{
\link[jjmR:readJJM]{jjmR::readJJM} \link[FLCore:FLBiol]{FLCore::FLBiol} buildFLBjjm
}
