% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildFLSsjjm.R, R/build.R
\name{buildFLSsjjm}
\alias{buildFLSsjjm}
\title{Build Multiple FLStock Objects from a jjm.output Object}
\usage{
buildFLSsjjm(out)

buildFLSsjjm(out)
}
\arguments{
\item{out}{A jjm.output object, typically the output of a stock assessment model.}
}
\value{
An FLStocks object containing multiple FLStock objects.

An FLStocks object containing multiple FLStock objects.
}
\description{
This function creates a collection of FLStock objects from a jjm.output object.
It is particularly useful for multi-stock assessments.

This function creates a collection of FLStock objects from a jjm.output object.
It is particularly useful for multi-stock assessments.
}
\seealso{
\code{\link[FLCore]{FLStock}}, \code{\link[FLCore]{FLStocks}}

\code{\link[FLCore]{FLStock}}, \code{\link[FLCore]{FLStocks}}
}
