% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{cjm.oem}
\alias{cjm.oem}
\title{Update Observations in CJM Fishery Operation Model}
\usage{
cjm.oem(stk, deviances, observations, args, tracking, F3sel)
}
\arguments{
\item{stk}{Stock data.}

\item{deviances}{Deviances for the model.}

\item{observations}{Observations data.}

\item{args}{Argument list containing model parameters.}

\item{tracking}{Tracking information.}

\item{F3sel}{F3 selectivity data.}
}
\value{
A list with updated stock data, indices, observations, and tracking information.
}
\description{
Updates stock and index observations in a CJM fishery operation model based on
provided arguments, tracking information, and F3 selectivity.
}
