% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readFLFsjjm}
\alias{readFLFsjjm}
\title{Create an FLFisheries from a JJMS model run directory}
\usage{
readFLFsjjm(name, path)
}
\arguments{
\item{name}{Name of the ctl model file, \emph{character}.}

\item{path}{Path ot the model folder structure, \emph{character}.}
}
\value{
An object of class FLIndices.
}
\description{
An object of class \emph{FLFisheries}, a list of \emph{FLFishery} objects, is created
from the information in the various files in a \emph{jjms} folder structure.
See \emph{buildFLFsjjm} for how each slot is populated from the inputs and
outputs of \emph{jjms}.
}
\examples{
fisheries <- readFLFsjjm(name="h1_1.07",
  path = system.file("ext-data", "single_stock", package="FLjjm"))
summary(fisheries)
}
\seealso{
\link[jjmR:readJJM]{jjmR::readJJM} \link[FLFishery:FLFisheries]{FLFishery::FLFisheries} \link[FLFishery:FLFishery]{FLFishery::FLFishery} buildFLFsjjm
}
