% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readFLIsjjm}
\alias{readFLIsjjm}
\title{Create an FLIndices from a JJMS model run directory}
\usage{
readFLIsjjm(name, path)
}
\arguments{
\item{name}{Name of the ctl model file, \emph{character}.}

\item{path}{Path ot the model folder structure, \emph{character}.}
}
\value{
An object of class FLIndices.
}
\description{
An object of class \emph{FLIndices}, a list of \emph{FLIndex} objects, is created
from the information in the various files in a \emph{jjms} folder structure.
See \emph{buildFLIsjjm} for how each slot is populated from the inputs and
outputs of \emph{jjms}.
}
\examples{
indices <- readFLIsjjm(name="mod1.00",
  path = system.file("ext-data", "single_stock", package="FLjjm"))
summary(indices)
}
\seealso{
\link[jjmR:readJJM]{jjmR::readJJM} \link[FLCore:FLIndices]{FLCore::FLIndices} \link[FLCore:FLIndex]{FLCore::FLIndex} buildFLIsjjm
}
