% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readFLoemjjm}
\alias{readFLoemjjm}
\title{Create an FLoem from a JJMS model run directory}
\usage{
readFLoemjjm(name, path, method = cjm.oem, iter = 1, ...)
}
\arguments{
\item{name}{Name of the ctl model file, \emph{character}.}

\item{path}{Path to the model folder structure, \emph{character}.}
}
\value{
An object of class FLoem.
}
\description{
An object of class \emph{FLoem} is created from the information in the various
files in a \emph{jjms} folder structure. The \emph{@observations} slot is created
containing the past observations of catch and biology (as an \emph{FLStock} or
\emph{FLStocks}) and for the indices of abundance (as an \emph{FLIndices})
}
\examples{
# One stock OM
oem <- readFLoemjjm(name="mod1.00",
  path=system.file("ext-data", "single_stock", package="FLjjm"))
oem
# Two stock OM
oemtwo <- readFLoemjjm(name="mod1.00_2stk",
  path=system.file("ext-data", "two_stock", package="FLjjm"))
oemtwo
}
\seealso{
\link[jjmR:readJJM]{jjmR::readJJM} \link[mse:FLoem-class]{mse::FLoem} buildFLSjjm buildFLIsjjm
}
