% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readFLomjjm}
\alias{readFLomjjm}
\title{Create an FLombf from a JJMS model run directory}
\usage{
readFLomjjm(name, path, iter = NULL, ...)
}
\arguments{
\item{name}{Name of the ctl model file, \emph{character}.}

\item{path}{Path to the model folder structure, \emph{character}.}
}
\value{
An object of class FLBiol.
}
\description{
An object of class \emph{FLombf} is created from the information in the various
files in a \emph{jjms} folder structure. Three slots are created from the
corresponding calls to \emph{buildFLjjm} functions.
\itemize{
\item @biols, of class \emph{FLBiols} from a call to \emph{buildFLBsjjm}.
\item @fisheries, of class \emph{FLFisheries}, from a call to \emph{buildFLFsjjm}.
\item @refpts, of classs \emph{FLPars}, from a call to \emph{buildFLRPsjjm}.
}
}
\examples{
# One stock OM
om <- readFLomjjm(name="h1_1.07",
  path=system.file("ext-data", "single_stock", package="FLjjm"))
summary(om)
# Two stock OM
omtwo <- readFLomjjm(name="h2_1.07",
  path=system.file("ext-data", "two_stock", package="FLjjm"))
summary(omtwo)
}
\seealso{
\link[jjmR:readJJM]{jjmR::readJJM} FLombf buildFLBjjm buildFLFsjjm
}
